Implementing Bill's changes to cuda driver so internal calculations are double precision

+++ cuda_driver/rx_cuda.cu
@@ -163,13 +163,13 @@ __global__ void multiply_mix_add(int16_t *samples, float *odata, float *filter)
     if (iSample_rf >= offset) {
         uint32_t idxSample_rf = (iSample_rf - offset) * 2 + iAntenna *  nSamples_rf *2;         // index in memory (account for  I/Q, iAntenna)
 
-        double phi0 = fmod(phaseIncrement_NCO_rad[iChannel] * iSample_rf, 2*M_PI);
-	float mc0=cos(phi0);
-	float ms0=sin(phi0);
+        double phi0 = fmod((double)phaseIncrement_NCO_rad[iChannel] * (double)iSample_rf, (double)(2*M_PI));
+        float mc0=cos(phi0);
+        float ms0=sin(phi0);
 
-        double phi1 = fmod(phaseIncrement_NCO_rad[iChannel] * (iSample_rf+1), 2*M_PI);
-	float mc1=cos(phi1);
-	float ms1=sin(phi1);
+        double phi1 = fmod((double)phaseIncrement_NCO_rad[iChannel] * (double)(iSample_rf+1), (double)(2*M_PI));
+        float mc1=cos(phi1);
+        float ms1=sin(phi1);
 
         itemp[iThread_lin] =
             filter[idxSample_filter  ] * samples[idxSample_rf  ] * mc0 -
